# packages/contracts/openapi/economics.yaml
openapi: 3.1.0
info:
  title: RAV5 Economic Service
  description: Economic calculation service for RAV5 robotics assessments.
  version: 0.1.0
paths:
...
  /api/v1/evaluations:
    post:
      tags:
      - evaluations
      summary: Create Evaluation
      description: 'Calculates, persists, and returns one evaluation snapshot.


        dry_run calculates without saving: the guest preview of a demo project

        keeps nothing (roles model §5).'
      operationId: create_evaluation_api_v1_evaluations_post
      parameters:
      - name: dry_run
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Dry Run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationRequestDto-Input'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationSnapshotDto'
        '409':
          description: The requested model version is unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '422':
          description: The request is invalid or contains invalid domain data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
  /api/v1/evaluations/{evaluation_id}:
    get:
      tags:
      - evaluations
      summary: Get Evaluation
      description: Returns the latest immutable snapshot for an evaluation.
      operationId: get_evaluation_api_v1_evaluations__evaluation_id__get
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationSnapshotDto'
        '404':
          description: The evaluation was not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    AcquisitionModel:
      type: string
      enum:
      - purchase
      - raas
      title: AcquisitionModel
      description: Supported candidate acquisition models.
...
    CandidateDto-Input:
      properties:
        candidate_id:
          type: string
          title: Candidate Id
        robot_code:
          type: string
          title: Robot Code
        price:
          anyOf:
          - $ref: '#/components/schemas/MoneyDto-Input'
          - type: 'null'
        payload_kg:
          anyOf:
          - type: number
          - type: string
          - type: 'null'
          title: Payload Kg
        max_speed_mps:
          anyOf:
          - type: number
          - type: string
          - type: 'null'
          title: Max Speed Mps
        loading_seconds:
          anyOf:
          - type: number
          - type: string
          - type: 'null'
          title: Loading Seconds
        unloading_seconds:
          anyOf:
          - type: number
          - type: string
          - type: 'null'
          title: Unloading Seconds
        average_power_kw:
          anyOf:
          - type: number
          - type: string
          - type: 'null'
          title: Average Power Kw
        handling_method:
          anyOf:
          - type: string
          - type: 'null'
          title: Handling Method
        catalog_status:
          type: string
          title: Catalog Status
        confirmation:
          $ref: '#/components/schemas/ConfirmationStatus'
        source:
          $ref: '#/components/schemas/SourceDto'
        maturity_trl:
          anyOf:
          - $ref: '#/components/schemas/SourcedValueDto-Input'
          - type: 'null'
        catalog_completeness_percent:
          anyOf:
          - $ref: '#/components/schemas/SourcedValueDto-Input'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - candidate_id
      - robot_code
      - price
      - payload_kg
      - max_speed_mps
      - loading_seconds
      - unloading_seconds
      - average_power_kw
      - handling_method
      - catalog_status
      - confirmation
      - source
      title: CandidateDto
...
    ConfirmationStatus:
      type: string
      enum:
      - confirmed
      - partial
      - unconfirmed
      - unknown
      title: ConfirmationStatus
      description: Describes whether a source value has been confirmed.
    ErrorResponseDto:
      properties:
        detail:
          anyOf:
          - type: string
          - items:
              $ref: '#/components/schemas/ValidationIssueDto'
            type: array
          title: Detail
      additionalProperties: false
      type: object
      required:
      - detail
      title: ErrorResponseDto
      description: Describes an HTTP error response body.
    EvaluationRequestDto-Input:
      properties:
        evaluation_id:
          type: string
          title: Evaluation Id
        project_id:
          type: string
          title: Project Id
        model_version:
          type: string
          title: Model Version
        task:
          $ref: '#/components/schemas/TaskDto-Input'
        candidates:
          items:
            $ref: '#/components/schemas/CandidateDto-Input'
          type: array
          title: Candidates
        norms:
          $ref: '#/components/schemas/NormsDto-Input'
        scenarios:
          items:
            $ref: '#/components/schemas/ScenarioDto-Input'
          type: array
          title: Scenarios
        requested_ranking_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Requested Ranking Version
        calculation_currency:
          type: string
          title: Calculation Currency
          default: RUB
        ranking_weights:
          anyOf:
          - items:
              prefixItems:
              - $ref: '#/components/schemas/RankingCriterion'
              - anyOf:
                - type: number
                - type: string
              type: array
              maxItems: 2
              minItems: 2
            type: array
          - type: 'null'
          title: Ranking Weights
      additionalProperties: false
      type: object
      required:
      - evaluation_id
      - project_id
      - model_version
      - task
      - candidates
      - norms
      - scenarios
      title: EvaluationRequestDto
    EvaluationRequestDto-Output:
      properties:
        evaluation_id:
          type: string
          title: Evaluation Id
        project_id:
          type: string
          title: Project Id
        model_version:
          type: string
          title: Model Version
        task:
          $ref: '#/components/schemas/TaskDto-Output'
        candidates:
          items:
            $ref: '#/components/schemas/CandidateDto-Output'
          type: array
          title: Candidates
        norms:
          $ref: '#/components/schemas/NormsDto-Output'
        scenarios:
          items:
            $ref: '#/components/schemas/ScenarioDto-Output'
          type: array
          title: Scenarios
        requested_ranking_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Requested Ranking Version
        calculation_currency:
          type: string
          title: Calculation Currency
          default: RUB
        ranking_weights:
          anyOf:
          - items:
              prefixItems:
              - $ref: '#/components/schemas/RankingCriterion'
              - type: string
              type: array
              maxItems: 2
              minItems: 2
            type: array
          - type: 'null'
          title: Ranking Weights
      additionalProperties: false
      type: object
      required:
      - evaluation_id
      - project_id
      - model_version
      - task
      - candidates
      - norms
      - scenarios
      title: EvaluationRequestDto
    EvaluationResultDto:
      properties:
        evaluation_id:
          type: string
          title: Evaluation Id
        project_id:
          type: string
          title: Project Id
        model_version:
          type: string
          title: Model Version
        status:
          $ref: '#/components/schemas/EvaluationStatus'
        candidates:
          items:
            $ref: '#/components/schemas/CandidateEconomicsDto'
          type: array
          title: Candidates
        ranking:
          $ref: '#/components/schemas/RankingResultDto'
      additionalProperties: false
      type: object
      required:
      - evaluation_id
      - project_id
      - model_version
      - status
      - candidates
      - ranking
      title: EvaluationResultDto
    EvaluationSnapshotDto:
      properties:
        evaluation_id:
          type: string
          title: Evaluation Id
        project_id:
          type: string
          title: Project Id
        request:
          $ref: '#/components/schemas/EvaluationRequestDto-Output'
        result:
          $ref: '#/components/schemas/EvaluationResultDto'
        created_at:
          type: string
          title: Created At
        revision_of:
          anyOf:
          - type: string
          - type: 'null'
          title: Revision Of
      additionalProperties: false
      type: object
      required:
      - evaluation_id
      - project_id
      - request
      - result
      - created_at
      - revision_of
      title: EvaluationSnapshotDto
    EvaluationStatus:
      type: string
      enum:
      - completed
      - partial
      - invalid
      title: EvaluationStatus
      description: Overall status of an evaluation.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MoneyDto-Input:
      properties:
        amount:
          anyOf:
          - type: number
          - type: string
          title: Amount
        currency:
          type: string
          title: Currency
        scale:
          type: integer
          title: Scale
          default: 0
      additionalProperties: false
      type: object
      required:
      - amount
      - currency
      title: MoneyDto
...
    NormsDto-Input:
      properties:
        payroll_multiplier:
          anyOf:
          - type: number
          - type: string
          title: Payroll Multiplier
        productive_time_share:
          anyOf:
          - type: number
          - type: string
          title: Productive Time Share
        technical_availability:
          anyOf:
          - type: number
          - type: string
          title: Technical Availability
        fleet_reserve_share:
          anyOf:
          - type: number
          - type: string
          title: Fleet Reserve Share
        operating_speed_factor:
          anyOf:
          - type: number
          - type: string
          title: Operating Speed Factor
        robots_per_charger:
          anyOf:
          - type: number
          - type: string
          title: Robots Per Charger
        charger_installed_price:
          $ref: '#/components/schemas/MoneyDto-Input'
        charger_power_kw:
          anyOf:
          - type: number
          - type: string
          title: Charger Power Kw
        fms_upfront_share:
          anyOf:
          - type: number
          - type: string
          title: Fms Upfront Share
        delivery_share:
          anyOf:
          - type: number
          - type: string
          title: Delivery Share
        commissioning_share:
          anyOf:
          - type: number
          - type: string
          title: Commissioning Share
        training_cost:
          $ref: '#/components/schemas/MoneyDto-Input'
        capex_contingency_share:
          anyOf:
          - type: number
          - type: string
          title: Capex Contingency Share
        annual_service_share:
          anyOf:
          - type: number
          - type: string
          title: Annual Service Share
        annual_license_share:
          anyOf:
          - type: number
          - type: string
          title: Annual License Share
        annual_repair_share:
          anyOf:
          - type: number
          - type: string
          title: Annual Repair Share
        electricity_price:
          $ref: '#/components/schemas/MoneyDto-Input'
        battery_life_years:
          anyOf:
          - type: number
          - type: string
          title: Battery Life Years
        battery_replacement_share:
          anyOf:
          - type: number
          - type: string
          title: Battery Replacement Share
        annual_connectivity_cost:
          $ref: '#/components/schemas/MoneyDto-Input'
        equipment_life_years:
          anyOf:
          - type: number
          - type: string
          title: Equipment Life Years
        discount_rate:
          anyOf:
          - type: number
          - type: string
          title: Discount Rate
        loan_share:
          anyOf:
          - type: number
          - type: string
          title: Loan Share
        loan_interest_rate:
          anyOf:
          - type: number
          - type: string
          title: Loan Interest Rate
        loan_term_years:
          anyOf:
          - type: number
          - type: string
          title: Loan Term Years
        monthly_raas_share:
          anyOf:
          - type: number
          - type: string
          title: Monthly Raas Share
        raas_setup_share:
          anyOf:
          - type: number
          - type: string
          title: Raas Setup Share
        recruitment_months_salary:
          anyOf:
          - type: number
          - type: string
          title: Recruitment Months Salary
        good_payback_years:
          anyOf:
          - type: number
          - type: string
          title: Good Payback Years
        medium_payback_years:
          anyOf:
          - type: number
          - type: string
          title: Medium Payback Years
        site_preparation_share:
          anyOf:
          - type: number
          - type: string
          title: Site Preparation Share
        source:
          $ref: '#/components/schemas/SourceDto'
      additionalProperties: false
      type: object
      required:
      - payroll_multiplier
      - productive_time_share
      - technical_availability
      - fleet_reserve_share
      - operating_speed_factor
      - robots_per_charger
      - charger_installed_price
      - charger_power_kw
      - fms_upfront_share
      - delivery_share
      - commissioning_share
      - training_cost
      - capex_contingency_share
      - annual_service_share
      - annual_license_share
      - annual_repair_share
      - electricity_price
      - battery_life_years
      - battery_replacement_share
      - annual_connectivity_cost
      - equipment_life_years
      - discount_rate
      - loan_share
      - loan_interest_rate
      - loan_term_years
      - monthly_raas_share
      - raas_setup_share
      - recruitment_months_salary
      - good_payback_years
      - medium_payback_years
      - site_preparation_share
      - source
      title: NormsDto
...
    RankingCriterion:
      type: string
      enum:
      - payback
      - roi
      - annual_effect
      - budget_fit
      - tco_savings
      - maturity
      - data_quality
      - fleet_utilization
      title: RankingCriterion
      description: Criteria supported by the ranking methodology contract.
...
    RankingResultDto:
      properties:
        status:
          $ref: '#/components/schemas/RankingStatus'
        model_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Version
        items:
          items:
            $ref: '#/components/schemas/RankingItemDto'
          type: array
          title: Items
      additionalProperties: false
      type: object
      required:
      - status
      - model_version
      - items
      title: RankingResultDto
...
    ScenarioDto-Input:
      properties:
        acquisition_model:
          $ref: '#/components/schemas/AcquisitionModel'
        price_factor:
          anyOf:
          - type: number
          - type: string
          title: Price Factor
        volume_factor:
          anyOf:
          - type: number
          - type: string
          title: Volume Factor
        labor_factor:
          anyOf:
          - type: number
          - type: string
          title: Labor Factor
        model_version:
          type: string
          title: Model Version
      additionalProperties: false
      type: object
      required:
      - acquisition_model
      - price_factor
      - volume_factor
      - labor_factor
      - model_version
      title: ScenarioDto
    ScenarioDto-Output:
      properties:
        acquisition_model:
          $ref: '#/components/schemas/AcquisitionModel'
        price_factor:
          type: string
          title: Price Factor
        volume_factor:
          type: string
          title: Volume Factor
        labor_factor:
          type: string
          title: Labor Factor
        model_version:
          type: string
          title: Model Version
      additionalProperties: false
      type: object
      required:
      - acquisition_model
      - price_factor
      - volume_factor
      - labor_factor
      - model_version
      title: ScenarioDto
    SourceDto:
      properties:
        source:
          type: string
          title: Source
        origin:
          $ref: '#/components/schemas/InputOrigin'
        confirmation:
          $ref: '#/components/schemas/ConfirmationStatus'
        version:
          anyOf:
          - type: string
          - type: 'null'
          title: Version
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
      additionalProperties: false
      type: object
      required:
      - source
      - origin
      - confirmation
      title: SourceDto
    SourcedValueDto-Input:
      properties:
        value:
          anyOf:
          - type: number
          - type: string
          title: Value
        unit:
          type: string
          title: Unit
        source:
          $ref: '#/components/schemas/SourceDto'
      additionalProperties: false
      type: object
      required:
      - value
      - unit
      - source
      title: SourcedValueDto
...
    TaskDto-Input:
      properties:
        project_id:
          type: string
          title: Project Id
        task_id:
          type: string
          title: Task Id
        operations_per_day:
          anyOf:
          - type: number
          - type: string
          title: Operations Per Day
        peak_factor:
          anyOf:
          - type: number
          - type: string
          title: Peak Factor
        automatable_share:
          anyOf:
          - type: number
          - type: string
          title: Automatable Share
        operating_hours_per_day:
          anyOf:
          - type: number
          - type: string
          title: Operating Hours Per Day
        one_way_route_m:
          anyOf:
          - type: number
          - type: string
          title: One Way Route M
        site_speed_limit_mps:
          anyOf:
          - type: number
          - type: string
          title: Site Speed Limit Mps
        load_unit_mass_kg:
          anyOf:
          - type: number
          - type: string
          title: Load Unit Mass Kg
        load_is_divisible:
          type: boolean
          title: Load Is Divisible
        available_charging_power_kw:
          anyOf:
          - type: number
          - type: string
          title: Available Charging Power Kw
        target_fte:
          anyOf:
          - type: number
          - type: string
          title: Target Fte
        target_annual_payroll:
          $ref: '#/components/schemas/MoneyDto-Input'
        baseline_annual_opex:
          $ref: '#/components/schemas/MoneyDto-Input'
          description: Annual baseline payroll for the current process workforce; this is not total process
            OPEX.
        fleet_operators_per_shift:
          anyOf:
          - type: number
          - type: string
          title: Fleet Operators Per Shift
        shifts_per_day:
          anyOf:
          - type: number
          - type: string
          title: Shifts Per Day
        staff_time_loss_share:
          anyOf:
          - type: number
          - type: string
          title: Staff Time Loss Share
        fleet_operator_monthly_salary:
          $ref: '#/components/schemas/MoneyDto-Input'
        target_monthly_salary:
          $ref: '#/components/schemas/MoneyDto-Input'
        annual_staff_turnover:
          anyOf:
          - type: number
          - type: string
          title: Annual Staff Turnover
        annual_other_benefits:
          $ref: '#/components/schemas/MoneyDto-Input'
        replacement_by_handling:
          items:
            prefixItems:
            - type: string
            - anyOf:
              - type: number
              - type: string
            type: array
            maxItems: 2
            minItems: 2
          type: array
          title: Replacement By Handling
        horizon_years:
          type: integer
          title: Horizon Years
        budget:
          anyOf:
          - $ref: '#/components/schemas/MoneyDto-Input'
          - type: 'null'
        source:
          $ref: '#/components/schemas/SourceDto'
        lift_trip_share:
          anyOf:
          - type: number
          - type: string
          title: Lift Trip Share
          default: '0'
        one_way_lift_seconds:
          anyOf:
          - type: number
          - type: string
          title: One Way Lift Seconds
          default: '0'
        integration_cost:
          anyOf:
          - $ref: '#/components/schemas/MoneyDto-Input'
          - type: 'null'
        annual_consumables_per_robot:
          anyOf:
          - $ref: '#/components/schemas/MoneyDto-Input'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - project_id
      - task_id
      - operations_per_day
      - peak_factor
      - automatable_share
      - operating_hours_per_day
      - one_way_route_m
      - site_speed_limit_mps
      - load_unit_mass_kg
      - load_is_divisible
      - available_charging_power_kw
      - target_fte
      - target_annual_payroll
      - baseline_annual_opex
      - fleet_operators_per_shift
      - shifts_per_day
      - staff_time_loss_share
      - fleet_operator_monthly_salary
      - target_monthly_salary
      - annual_staff_turnover
      - annual_other_benefits
      - replacement_by_handling
      - horizon_years
      - budget
      - source
      title: TaskDto
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    ValidationIssueDto:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Loc
        msg:
          type: string
          title: Msg
        type:
          type: string
          title: Type
        input:
          anyOf:
          - {}
          - type: 'null'
          title: Input
        ctx:
          anyOf:
          - type: object
          - type: 'null'
          title: Ctx
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationIssueDto
      description: Describes one request validation issue returned by FastAPI.